requirements.txt
fastapi
uvicorn
twilio
python-multipart
gunicorn
pydantic
python-dotenv
axios
requests
google-cloud-speech
aiohttp
google-cloud-storage